// src/types/dto/response/employee-response.dto.ts
export interface EmployeeResponseDto {
  id: string;
  nome: string;
  telefone?: string;
  email?: string;
  rg?: string;
  cpf?: string;
  endereco?: string;
  cargo?: string;
  crefito?: string;
  filialId?: string;
  dataNascimento?: string;
  contatoEmergencial?: string;
  cor?: string;
  dataCadastro?: string;
  ativo?: boolean;
}